migrate blog to hugo
I was more familiar with Hugo so I migrated everything from Jekyll to
Hugo.

I also fixed all the blog posts so the images load and links work.

I themed the blog so it matches the main website as well.

I did not bother to get pagination working. Can do that later if we feel
like it. For now, things load!

Signed-off-by: Taylor Silva <dev@taydev.net>